Consolidated FormatDescriptor into VersionedSerializer.
